Update to the edits branch
This update includes correctons of my database structure so it is easier to handle the courses and their analysis,
by introducing the course offering and making the course branch it makes it easier to register same course code with different branching and units based on department.

+++ eduNalytics/collector/models.py
@@ -64,9 +64,7 @@ class CourseResult(models.Model):
         max_length=2,
         choices=[(chr(i), chr(i)) for i in range(ord('A'), ord('F')+1)]
     )
-    score = models.DecimalField(
-        max_digits=5, 
-        decimal_places=2,
+    score = models.PositiveIntegerField(
         validators=[MaxValueValidator(100)]
     )
